requirements.txt
arxiv==2.2.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
feedparser==6.0.11
Flask==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
psycopg==3.2.9
psycopg-binary==3.2.9
python-dotenv==1.1.0
requests==2.32.4
sgmllib3k==1.0.0
urllib3==2.4.0
Werkzeug==3.1.3
